--- a/benches/large_case.rs
+++ b/benches/large_case.rs
@@ -1,24 +1,38 @@
 // SPDX-License-Identifier: MPL-2.0
-
+use std::fmt::{Debug, Display};
+use std::hash::Hash;
 use std::time::Duration;
 
 use criterion::*;
+use pubgrub::{Map, OfflineDependencyProvider, Range, VersionRanges};
 use serde::de::Deserialize;
 
-use pubgrub::{resolve, OfflineDependencyProvider, Package, Range, SemanticVersion, VersionSet};
-
-fn bench<'a, P: Package + Deserialize<'a>, VS: VersionSet + Deserialize<'a>>(
+fn bench<
+    'a,
+    P: Debug + Display + Clone + Eq + Hash + Deserialize<'a>,
+    R: VersionRanges + Deserialize<'a>,
+>(
     b: &mut Bencher,
     case: &'a str,
 ) where
-    <VS as VersionSet>::V: Deserialize<'a>,
+    R::V: Deserialize<'a>,
 {
-    let dependency_provider: OfflineDependencyProvider<P, VS> = ron::de::from_str(case).unwrap();
+    let mut dependency_provider: OfflineDependencyProvider<P, R> = ron::de::from_str(case).unwrap();
+
+    let dependencies = dependency_provider
+        .packages()
+        .map(|p| {
+            (
+                p.clone(),
+                dependency_provider.versions(p).unwrap().cloned().collect(),
+            )
+        })
+        .collect::<Map<_, Vec<_>>>();
 
     b.iter(|| {
-        for p in dependency_provider.packages() {
-            for n in dependency_provider.versions(p).unwrap() {
-                let _ = resolve(&dependency_provider, p.clone(), n.clone());
+        for (p, versions) in &dependencies {
+            for v in versions {
+                let _ = dependency_provider.resolve(p.clone(), v.clone());
             }
         }
     });
@@ -32,13 +46,9 @@ fn bench_nested(c: &mut Criterion) {
         let case = case.unwrap().path();
         let name = case.file_name().unwrap().to_string_lossy();
         let data = std::fs::read_to_string(&case).unwrap();
-        if name.ends_with("u16_NumberVersion.ron") || name.ends_with("u16_u32.ron") {
+        if name.ends_with("u16_NumberVersion.ron") {
             group.bench_function(name, |b| {
                 bench::<u16, Range<u32>>(b, &data);
-            });
-        } else if name.ends_with("str_SemanticVersion.ron") {
-            group.bench_function(name, |b| {
-                bench::<&str, Range<SemanticVersion>>(b, &data);
             });
         }
     }