--- a/examples/branching_error_reporting.rs
+++ b/examples/branching_error_reporting.rs
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: MPL-2.0
 
 use pubgrub::{
-    resolve, DefaultStringReporter, OfflineDependencyProvider, PubGrubError, Ranges, Reporter,
+    DefaultStringReporter, OfflineDependencyProvider, PubGrubError, Ranges, Reporter,
     SemanticVersion,
 };
 
@@ -10,15 +10,16 @@ type SemVS = Ranges<SemanticVersion>;
 // https://github.com/dart-lang/pub/blob/master/doc/solver.md#branching-error-reporting
 fn main() {
     let mut dependency_provider = OfflineDependencyProvider::<&str, SemVS>::new();
+
     #[rustfmt::skip]
     // root 1.0.0 depends on foo ^1.0.0
-        dependency_provider.add_dependencies(
+    dependency_provider.add_dependencies(
         "root", (1, 0, 0),
         [("foo", Ranges::from_range_bounds((1, 0, 0)..(2, 0, 0)))],
     );
     #[rustfmt::skip]
     // foo 1.0.0 depends on a ^1.0.0 and b ^1.0.0
-        dependency_provider.add_dependencies(
+    dependency_provider.add_dependencies(
         "foo", (1, 0, 0),
         [
             ("a", Ranges::from_range_bounds((1, 0, 0)..(2, 0, 0))),
@@ -27,7 +28,7 @@ fn main() {
     );
     #[rustfmt::skip]
     // foo 1.1.0 depends on x ^1.0.0 and y ^1.0.0
-        dependency_provider.add_dependencies(
+    dependency_provider.add_dependencies(
         "foo", (1, 1, 0),
         [
             ("x", Ranges::from_range_bounds((1, 0, 0)..(2, 0, 0))),
@@ -36,7 +37,7 @@ fn main() {
     );
     #[rustfmt::skip]
     // a 1.0.0 depends on b ^2.0.0
-        dependency_provider.add_dependencies(
+    dependency_provider.add_dependencies(
         "a", (1, 0, 0),
         [("b", Ranges::from_range_bounds((2, 0, 0)..(3, 0, 0)))],
     );
@@ -45,7 +46,7 @@ fn main() {
     dependency_provider.add_dependencies("b", (2, 0, 0), []);
     #[rustfmt::skip]
     // x 1.0.0 depends on y ^2.0.0.
-        dependency_provider.add_dependencies(
+    dependency_provider.add_dependencies(
         "x", (1, 0, 0),
         [("y", Ranges::from_range_bounds((2, 0, 0)..(3, 0, 0)))],
     );
@@ -54,11 +55,14 @@ fn main() {
     dependency_provider.add_dependencies("y", (2, 0, 0), []);
 
     // Run the algorithm.
-    match resolve(&dependency_provider, "root", (1, 0, 0)) {
+    match dependency_provider.resolve("root", (1, 0, 0)) {
         Ok(sol) => println!("{:?}", sol),
-        Err(PubGrubError::NoSolution(mut derivation_tree)) => {
-            derivation_tree.collapse_no_versions();
-            eprintln!("{}", DefaultStringReporter::report(&derivation_tree));
+        Err(PubGrubError::NoSolution(mut error)) => {
+            error.derivation_tree.collapse_no_versions();
+            eprintln!(
+                "{}",
+                DefaultStringReporter::report(&error, &dependency_provider)
+            );
             std::process::exit(1);
         }
         Err(err) => panic!("{:?}", err),