--- a/examples/doc_interface.rs
+++ b/examples/doc_interface.rs
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: MPL-2.0
 
-use pubgrub::{resolve, OfflineDependencyProvider, Ranges};
+use pubgrub::{OfflineDependencyProvider, Ranges};
 
 type NumVS = Ranges<u32>;
 
@@ -19,6 +19,6 @@ fn main() {
     dependency_provider.add_dependencies("icons", 1u32, []);
 
     // Run the algorithm.
-    let solution = resolve(&dependency_provider, "root", 1u32);
-    println!("Solution: {:?}", solution);
+    let solution = dependency_provider.resolve("root", 1u32).unwrap();
+    println!("Solution: {solution:?}");
 }