--- a/src/error.rs
+++ b/src/error.rs
@@ -4,30 +4,34 @@
 
 use thiserror::Error;
 
-use crate::{DependencyProvider, DerivationTree};
+use crate::{DependencyProvider, DerivationTree, PackageArena};
 
 /// There is no solution for this set of dependencies.
-pub type NoSolutionError<DP> = DerivationTree<
-    <DP as DependencyProvider>::P,
-    <DP as DependencyProvider>::VS,
-    <DP as DependencyProvider>::M,
->;
+#[derive(Debug, Clone)]
+pub struct NoSolutionError<DP: DependencyProvider> {
+    /// Package store
+    pub package_store: PackageArena<DP::P>,
+    /// Derivate tree
+    pub derivation_tree: DerivationTree<DP::M>,
+}
 
 /// Errors that may occur while solving dependencies.
 #[derive(Error)]
 pub enum PubGrubError<DP: DependencyProvider> {
+    /// Root package name doesn't exist.
+    #[error("Root package name doesn't exist")]
+    NoRoot,
+
     /// There is no solution for this set of dependencies.
     #[error("No solution")]
-    NoSolution(NoSolutionError<DP>),
+    NoSolution(Box<NoSolutionError<DP>>),
 
     /// Error arising when the implementer of [DependencyProvider] returned an error in the method
     /// [get_dependencies](DependencyProvider::get_dependencies).
-    #[error("Retrieving dependencies of {package} {version} failed")]
+    #[error("Retrieving dependencies of {package_version} failed")]
     ErrorRetrievingDependencies {
-        /// Package whose dependencies we want.
-        package: DP::P,
-        /// Version of the package for which we want the dependencies.
-        version: DP::V,
+        /// Represenatation of package and version whose dependencies we want.
+        package_version: String,
         /// Error raised by the implementer of
         /// [DependencyProvider].
         source: DP::Err,
@@ -50,7 +54,7 @@ pub enum PubGrubError<DP: DependencyProvider> {
 
 impl<DP: DependencyProvider> From<NoSolutionError<DP>> for PubGrubError<DP> {
     fn from(err: NoSolutionError<DP>) -> Self {
-        Self::NoSolution(err)
+        Self::NoSolution(err.into())
     }
 }
 
@@ -60,15 +64,18 @@ where
 {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
-            Self::NoSolution(err) => f.debug_tuple("NoSolution").field(&err).finish(),
+            Self::NoRoot => f.debug_struct("NoRoot").finish(),
+            Self::NoSolution(err) => f
+                .debug_struct("NoSolution")
+                .field("package_store", &err.package_store)
+                .field("derivation_tree", &err.derivation_tree)
+                .finish(),
             Self::ErrorRetrievingDependencies {
-                package,
-                version,
+                package_version,
                 source,
             } => f
                 .debug_struct("ErrorRetrievingDependencies")
-                .field("package", package)
-                .field("version", version)
+                .field("package_version", package_version)
                 .field("source", source)
                 .finish(),
             Self::ErrorChoosingPackageVersion(arg0) => f