--- a/src/internal/mod.rs
+++ b/src/internal/mod.rs
@@ -2,16 +2,12 @@
 
 //! Non exposed modules.
 
-mod arena;
 mod core;
 mod incompatibility;
 mod partial_solution;
 mod small_map;
-mod small_vec;
 
-pub(crate) use arena::{Arena, HashArena, Id};
 pub(crate) use core::State;
-pub(crate) use incompatibility::{IncompDpId, IncompId, Incompatibility, Relation};
+pub(crate) use incompatibility::{IncompatArena, IncompatId, Incompatibility, Relation};
 pub(crate) use partial_solution::{DecisionLevel, PartialSolution, SatisfierSearch};
 pub(crate) use small_map::SmallMap;
-pub(crate) use small_vec::SmallVec;