… Cargo.lock
… Cargo.toml
… README.md
… examples/caching_dependency_provider.rs
… examples/doc_interface_error.rs
… examples/doc_interface_semantic.rs
… examples/linear_error_reporting.rs
… examples/many_versions.rs
… examples/unsat_root_message_no_version.rs
… src/internal/arena.rs
… src/internal/core.rs
… src/internal/incompatibility.rs
… src/internal/partial_solution.rs
… src/internal/small_vec.rs
… src/lib.rs
… src/provider.rs
… src/report.rs
… src/semantic.rs
… src/solver.rs
… src/term.rs
… src/version.rs
… src/version_set.rs
… tests/examples.rs
… tests/proptest.rs
… tests/sat_dependency_provider.rs
… tests/tests.rs