This PR is an experiment to improve performance of Pubgrub.
There are many API breaking changes which should be reviewed before this can be merged.

The following benchmarks are done using the new bench file `bench_backtracking.ron`:

| Commit                                                  | Duration         |
|---------------------------------------------------------|------------------|
| `dev` branch                                            | 6.5s (estimated) |
| Use an arena for packages                               | 7.58s            |
| Use u64 for version set type                            | 1.21s            |
| Optimize incompatibility contradiction                  | 661ms            |
| Reduce AssignmentsIntersection size                     | 660ms            |
| Use smallvec crate in pubgrub                           | 604ms            |
| Use a Vec for incompatibilities and package assignments | 480ms            |

Results for the `large_case` bench:

| Commit         | Duration |
|----------------|----------|
| `dev` branch   | 5.0616ms |
| This PR        | 3.5688ms |

Implementation example for `cargo`: https://github.com/x-hgg-x/pubgrub-bench