@Eh2406 2024-11-07
Those are very impressive numbers. Thank you for looking into this!

I appreciate getting to see the unified vision of where everything is going. Obviously, with this much going on we are not going to be able to review it all in one PR. It's up to you how you want to divide  things up into reviewable pieces.

@konstin just added some really cool automation for running our benchmarks in https://github.com/pubgrub-rs/pubgrub/pull/271,  which apparently I forgot to click merge on. So if we get the benchmark merged first  we can easily and automatically see the improvements from the other pieces.

The non breaking changes do not seem all that controversial. So they may be  easy to land.  Of course, they may not pull their weight until the other (API-breaking) optimizations have come in.

The breaking changes will need to be looked at very carefully. Certainly each as their own PR. An order of magnitude performance improvement is nothing to sneeze at, but performance is not our only priority.  Our existing API  is careful to be "misuse resistant"  in that it is hard to  get an incorrect answer because you were "holding it wrong".

I look forward to reviewing this work, one piece at a time, over the next weeks/months.