@konstin 2024-11-08
That's an impressive speedup, and some deep changes!

Long and ramble-y comment ahead, here are some unstructured thoughts about this:

From spending a lot of time optimizing uv, i found that often making pubgrub do less work is the most effective strategy, for example by package prioritization (https://github.com/astral-sh/uv/blob/14507a1793f12bf884786be49d716ca8f8322340/crates/uv-resolver/src/pubgrub/priority.rs#L25) or https://github.com/astral-sh/uv/issues/8157. If we're checking a large (exponential) combination of packages of versions, trying to get to a package selection that decides things that are either easy/clear or that reduce the remaining space a lot avoids doing work in pubgrub.

How did you create `bench_backtracking.ron`? Is there a way to bench the changes with cargo?

By splitting out the non-breaking changes, we can bench them in uv.

We need to decide which optimizations are in pubgrub and which are outside. For example, in uv we `Arc` our version and package types, so cloning them in pubgrub is cheap. When https://github.com/pubgrub-rs/pubgrub/pull/242 or the arena-for-packages commit in this PR lands, we may stop `Arc`ing our package type, and instead have pubgrub do that, since `Arc` inside an arena would be duplicated.

I'm not following along with the new `Version` type, how do i pass in a `pep440_rs` version with the new type? How do i represent large versions?

What version type is cargo currently using, can we do an optimization for small versions? One thing a had pitched to @Eh2406 earlier: `semver::Version` is 40 bytes, while most versions fit in a `(u8, u8, u16)` for `(major, minor, patch)`, which can be in an enum with `Box<semver::Version>` for a total `u64` type. Version comparisons are then a u32 comparison for most versions. We do similar optimizations in https://github.com/astral-sh/uv/blob/a80b6f51b07e0706800a0a9e8e0f5fe42651aa1b/crates/uv-pep440/src/version.rs#L263-L282.

`Remove Eq bound for M`: We want to retain that bound for https://github.com/pubgrub-rs/pubgrub/issues/232.