@x-hgg-x 2024-11-08
> How did you create `bench_backtracking.ron`? Is there a way to bench the changes with cargo?

This file was created from cargo, using https://github.com/x-hgg-x/pubgrub-bench, by dumping the output of `get_dependencies()` for `solana-archiver-lib v1.1.12`.

> By splitting out the non-breaking changes, we can bench them in uv.

I will make separate PRs for each commit, this PR only shows the global performance impact.

> I'm not following along with the new `Version` type, how do i pass in a `pep440_rs` version with the new type? How do i represent large versions?

`Version` is an index of the true version (`semver::Version` for cargo), and is limited to 63 versions so that `VersionSet` and `Term` can be represented as a `u64` bitset, making intersection and union essentially free (this is the main improvement of this PR).

To use it, a dependency provider will need to store a list of all versions for a given package, which should be sorted by priority so that `choose_version()` can always choose the last `Version` of a `VersionSet`.

To allow for more than 63 versions of a package, virtual packages need to be defined to ensure package unicity and to allow more than 63 possible versions of a dependency:
* see https://github.com/x-hgg-x/pubgrub/blob/perf/src/helpers.rs for the definition of a wrapper to handle this (this creates one virtual package for each power of 63 versions).
* see https://github.com/x-hgg-x/pubgrub/blob/perf/examples/many_versions.rs for a simple example.
* see https://github.com/x-hgg-x/pubgrub-bench/blob/2037b78c0c90e9b2c14d7881b532b6e1c2cb74ef/src/provider.rs#L1162-L1190 for the cargo implementation.

Note that this complexifies version choosing, but we can avoid difficulties by sorting versions by priority in the dependency provider.