@mpizenberg 2024-11-08
This is an impressive speedup. I don’t have the bandwidth personally to review this unfortunately so I’ll just trust the other pairs of eyes already on this. Few remarks, I agree with the others about the possibilities to split this between non-controversial changes and more controversial changes. But it seems you already agree with that too so all good.

Another broad remark, I feel like some of the speedups there rely on specific representations for versions if I understand correctly what you did with the u64 bitset. I’d just want to point out the fact that it’s important to keep the overall API generic so that version sets details of implementation can stay user-specific and not embedded directly into pubgrub assumptions.

Nothing else to add except thank you for all that work! and I’m excited to (sporadically) follow along.