@x-hgg-x 2024-11-08
> Another broad remark, I feel like some of the speedups there rely on specific representations for versions if I understand correctly what you did with the u64 bitset. 

Dependency providers can still internally use whatever version format the want, but with this PR Pubgrub only accepts discrete versions for a dependency in the `get_dependencies()` method, which are representing the indices of provider versions.

This means we need to list all versions of a dependency matching the requirement in the `get_dependencies()` method, then give Pubgrub the corresponding bitset.

The `VersionSet` trait is therefore not needed anymore since Pubgrub will do the intersections/unions on the bitset (instead of on the range corresponding to the provider version). This is why the API doesn't need to be generic anymore.