@x-hgg-x 2024-11-09
> `Version` is an index of the true version

I will rename `Version` to `VersionIndex` to clarify this.