@Eh2406 2024-11-19
Those are very exciting numbers! What piece do you want to work on next? The first of your commits is https://github.com/pubgrub-rs/pubgrub/pull/274/commits/58340832df9029c8e239af18b1e4bd4f054af5e9 `Move offline provider into a separate module` which I've long wanted to do. So that is an ez +1. The next is https://github.com/pubgrub-rs/pubgrub/pull/274/commits/eeb525b57421274c658191b546c0202bd2d9441a `Use an arena for packages`,  which has a lot going on but I'm very supportive of the effort. #242 is my attempt to do something similar without breaking changes. If you could look that over  I'd appreciate it. Specifically any optimizations from your work that  should be included in that PR? Generally, I want to continue working for the non-breaking changes  first.