@x-hgg-x 2024-11-27
I added a method in the dependency provider to register a conflict when backtracking for better package prioritization, and it shows good results for `solana-archiver-lib v1.1.12`:

Before:
```
should cancel count: 7429904
backtracks: 8592
incompatibilities: 424401
packages: 6802
time for unit propagation: 1.138949313s
time for pick highest priority pkg: 574.226395ms
```

After:
```
should cancel count: 274934
backtracks: 8597
incompatibilities: 407651
packages: 6441
time for unit propagation: 143.769512ms
time for pick highest priority pkg: 62.437494ms
```

Conflict map:
```
(7212, "solana-sdk")
(1212, "solana-metrics")
(966, "solana-clap-utils")
(545, "ed25519-dalek")
(501, "solana-client")
(482, "solana-vote-program")
(404, "solana-version")
(398, "solana-perf")
(395, "solana-net-utils")
(384, "solana-runtime")
(369, "solana-config-program")
(359, "solana-account-decoder")
(353, "solana-measure")
(333, "solana-faucet")
(317, "solana-ledger")
(296, "solana-core")
(282, "solana-transaction-status")
(278, "cc")
(269, "solana-stake-program")
(267, "solana-remote-wallet")
(266, "solana-bpf-loader-program")
(247, "solana-sys-tuner")
(150, "solana-crate-features")
(68, "solana-program")
(64, "solana-exchange-program")
(63, "curve25519-dalek")
(52, "solana-vote-signer")
(50, "openssl-sys")
(32, "solana-budget-program")
(30, "solana-vest-program")
(21, "cxx")
(16, "spl-token")
(9, "ring")
(8, "regex-automata")
(4, "aho-corasick")
(3, "openssl")
(2, "solana-sdk-macro")
(1, "solana-genesis-programs")
```